# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protomind API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/protocols": {
            "get": {
                "tags": [
                    "Protocols"
                ],
                "summary": "Показать список протоколов",
                "description": "Метод для отображения списка протоколов",
                "operationId": "indexProtocols",
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProtocolResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Protocols"
                ],
                "summary": "Создать новый протокол",
                "description": "Метод для создания нового протокола",
                "operationId": "storeProtocol",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProtocolStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProtocolResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибочный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка при создании протокола"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/protocols/{id}": {
            "get": {
                "tags": [
                    "Protocols"
                ],
                "summary": "Показать протокол",
                "description": "Метод для отображения данных протокола",
                "operationId": "showProtocol",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProtocolResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Протокол не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Protocols"
                ],
                "summary": "Обновить протокол",
                "description": "Метод для обновления данных протокола",
                "operationId": "updateProtocol",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProtocolUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProtocolResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Протокол не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Protocols"
                ],
                "summary": "Удалить протокол",
                "description": "Метод для удаления протокола",
                "operationId": "destroyProtocol",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Протокол не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProtocolStoreRequest": {
                "title": "Запрос на создание протокола",
                "description": "Запрос, содержащий данные для создания протокола",
                "properties": {
                    "theme": {
                        "description": "Тема протокола",
                        "type": "string",
                        "maxLength": 255
                    },
                    "agenda": {
                        "description": "Повестка дня",
                        "type": "string",
                        "maxLength": 255
                    },
                    "secretary_id": {
                        "description": "ID секретаря",
                        "type": "integer"
                    },
                    "director_id": {
                        "description": "ID директора",
                        "type": "integer"
                    },
                    "event_date": {
                        "description": "Дата события",
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-01"
                    },
                    "video": {
                        "description": "Видео файл",
                        "type": "string",
                        "format": "binary",
                        "example": "video.mp4"
                    }
                },
                "type": "object"
            },
            "ProtocolUpdateRequest": {
                "title": "Запрос на обновление протокола",
                "description": "Запрос, содержащий данные для обновления протокола",
                "properties": {
                    "theme": {
                        "description": "Тема протокола",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "agenda": {
                        "description": "Повестка дня",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "secretary_id": {
                        "description": "ID секретаря",
                        "type": "integer",
                        "nullable": true
                    },
                    "director_id": {
                        "description": "ID директора",
                        "type": "integer",
                        "nullable": true
                    },
                    "event_date": {
                        "description": "Дата события",
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-01",
                        "nullable": true
                    },
                    "video": {
                        "description": "Видео файл",
                        "type": "string",
                        "format": "binary",
                        "example": "video.mp4",
                        "nullable": true
                    },
                    "final_transcript": {
                        "description": "Окончательный протокол",
                        "type": "string",
                        "nullable": true
                    },
                    "execute": {
                        "description": "Выполнить",
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProtocolResource": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Protocol"
                    }
                ]
            },
...
            "Protocol": {
                "title": "Протокол",
                "description": "Модель протокола",
                "properties": {
                    "id": {
                        "description": "ID протокола",
                        "type": "integer"
                    },
                    "theme": {
                        "description": "Тема протокола",
                        "type": "string"
                    },
                    "agenda": {
                        "description": "Повестка дня",
                        "type": "string"
                    },
                    "secretary_id": {
                        "description": "ID секретаря",
                        "type": "integer"
                    },
                    "director_id": {
                        "description": "ID директора",
                        "type": "integer"
                    },
                    "event_date": {
                        "description": "Дата события",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "description": "Статус протокола",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Дата и время создания протокола",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления протокола",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}